server/requirements.txt
eventlet==0.26.1
Flask==1.1.2
Flask-SocketIO==4.2.1
python-socketio==4.3
python-engineio==3.13.2
numpy==1.18.4
pyserial==3.4
scipy==1.5.4
tinydb==3.15.2
